Only provide hint and do not play automatically when suggesting move
