+++ mobile/src/App.css
@@ -46,7 +46,6 @@ main {
   width: 100%;
   display: flex;
   flex-direction: column;
-  justify-content: center;
   align-items: center;
   background-color: #f0f0f0;
   gap: 20px;
