+++ mobile/src/board/components/Tile.css
@@ -1,3 +1,28 @@
+@keyframes horizontal-shaking {
+  0% {
+    transform: translateX(0) translateY(0);
+  }
+  20% {
+    transform: translateX(-1px) translateY(-1px);
+  }
+  40% {
+    transform: translateX(1px) translateY(-1px);
+  }
+  60% {
+    transform: translateX(1px) translateY(1px);
+  }
+  80% {
+    transform: translateX(-1px) translateY(1px);
+  }
+  100% {
+    transform: translateX(0) translateY(0);
+  }
+}
+
+.hint {
+  animation: horizontal-shaking 0.5s ease-out infinite;
+}
+
 .tile {
   position: relative;
   display: flex;
@@ -18,8 +43,8 @@
 }
 
 .tile--remaining {
-  transition: transform 0.25s ease-out, opacity 0.5s ease-out;
   position: absolute;
+  transition: transform 0.25s ease-out, opacity 0.25s ease-out;
 }
 
 .tile--animate {
