--- mobile/src/board/hooks/index.ts
@@ -1,3 +1,0 @@
-export { useBoard } from "./useBoard";
-export { useNewBoardMutation } from "./useNewBoardMutation";
-export { useGetPlaceTileHintMutation } from "./useGetPlaceTileHintMutation";
