+++ mobile/src/board/hooks/useGetPlaceTileHintMutation.ts
@@ -1,21 +1,17 @@
 import { AlertColor } from "@mui/material";
 import { useMutation } from "react-query";
 
-import { type Board, type Error, type PlaceTileHint } from "../BoardTypes";
+import type { BoardID, Error, PlaceTileHint } from "../BoardTypes";
 
 type MutationError = { message: string; severity?: AlertColor };
 
-interface PlaceTileHintResponseData {
-  placeTileHint: PlaceTileHint | null;
-}
-
 interface PlaceTileHintResponse {
-  data: MutationError | PlaceTileHintResponseData;
+  data: MutationError | PlaceTileHint;
 }
 
 export const useGetPlaceTileHintMutation = () =>
-  useMutation<PlaceTileHint | null, Error, Board, void>(async (board) => {
-    const response = await fetch(`/api/v1/board/${board.id}/place-tile-hint`, {
+  useMutation<PlaceTileHint | null, Error, BoardID, void>(async (id) => {
+    const response = await fetch(`/api/v1/board/${id}/place-tile-hint`, {
       method: "PUT",
       headers: {
         "Content-Type": "application/json",
@@ -31,9 +27,10 @@ export const useGetPlaceTileHintMutation = () =>
       };
     }
 
-    const data = responseContent.data as PlaceTileHintResponseData;
+    const data = responseContent.data as PlaceTileHint;
     if (!data) {
       return null;
     }
-    return data.placeTileHint;
+
+    return data;
   });
